updte

+++ models/stagging/raw/stg_raw__ship.sql
@@ -2,5 +2,5 @@ SELECT
     orders_id
     ,shipping_fee
     ,logCost
-    , CAST(ship_cost AS INT64) AS ship_cost
+    , CAST(ship_cost AS FLOAT64) AS ship_cost
 FROM {{ source('raw', 'ship') }}
